third_party/WebKit/LayoutTests/fast/forms/textarea/textarea-value-last-eol.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="log"></div>
<textarea></textarea>
<script>
test(function() {
    var textarea = document.querySelector('textarea');
    textarea.focus();
    eventSender.keyDown('H');
    eventSender.keyDown('\n');
    eventSender.keyDown('W');
    assert_equals(textarea.value, 'H\nW');
    textarea.setSelectionRange(2, 3); // "W"
    document.execCommand('cut');
    textarea.setSelectionRange(0, 0);
    document.execCommand('paste');
    // TODO(tkent): The following assertion fails now. crbug.com/522144.
    assert_equals(textarea.value, 'WH\n');
}, 'TEXTAREA should not remove the last EOL on paste.');
</script>
</body>
